requirements_prod.txt
# Core
python-telegram-bot==20.7
aiohttp==3.9.1
aiogram==3.3.0  # для webhook handler

# Database
SQLAlchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Task Queue
celery[redis]==5.3.4
redis==5.0.1
flower==2.0.1  # monitoring

# Video/Image Processing
ffmpeg-python==0.2.0
Pillow==10.1.0
yt-dlp==2023.12.30

# Utils
python-dotenv==1.0.0
pyotp==2.9.0
qrcode==7.4.2
psutil==5.9.6
aiofiles==23.2.1

# Monitoring
prometheus-client==0.19.0
structlog==23.2.0

# Performance
uvloop==0.19.0  # faster event loop
orjson==3.9.10  # faster JSON
